seo-generator/requirements.txt
psycopg2-binary>=2.9
Jinja2>=3.1
unidecode>=1.3
